feat(ticket): track agent usage costs and budgets
Summary:
- add parsed token and cost usage accounting to ticket service and expose
  token totals on ticket responses
- let agent tokens report usage through the platform API and CLI while
  extending default scopes for usage reporting
- switch metrics counters to float64 so usd totals can be emitted
  accurately and add focused coverage plus refreshed OpenAPI output

Rationale:
- F70 needs a narrow but end-to-end cost tracking slice that fits the
  current agent token and platform API flow
- keeping accounting in ticket service preserves a single write path for
  ticket cost fields, budget pause state, and metrics emission
- float counters avoid encoding usd totals as fake integers just to fit
  the metrics abstraction

Tests:
- PATH="$PWD/.tooling/go/bin:/home/yuzhong/.local/go1.26.1/bin:$PATH"
  go test ./internal/domain/ticketing ./internal/infra/otel
  ./internal/ticket -run
  'Test(ParseRawUsageDeltaRejectsEmptyPayload|UsageDeltaResolveCostUSDUsesExplicitCostWhenProvided|UsageDeltaResolveCostUSDComputesProviderPricing|ServiceRecordUsageAccumulatesTokensCostAndBudgetPause)$'
- PATH="$PWD/.tooling/go/bin:/home/yuzhong/.local/go1.26.1/bin:$PATH"
  go test ./internal/cli -run
  'Test(TicketReportUsageCommandPostsUsagePayload|TicketCreateCommandUsesAgentPlatformEnvironment|TicketUpdateCommandFallsBackToCurrentTicketEnv)$'
- PATH="$PWD/.tooling/go/bin:/home/yuzhong/.local/go1.26.1/bin:$PATH"
  go test ./internal/httpapi -run
  'Test(AgentPlatformTicketRoutesRespectScopesAndBoundaries|AgentPlatformHarnessWhitelistConstrainsTokenScopes)$'
- PATH="$PWD/.tooling/go/bin:/home/yuzhong/.local/go1.26.1/bin:$PATH"
  go test ./...
- ./scripts/ci/lint.sh

+++ internal/provider/metrics.go
@@ -5,7 +5,7 @@ type Tags map[string]string
 
 // Counter records additive totals.
 type Counter interface {
-	Add(int64)
+	Add(float64)
 }
 
 // Histogram records sampled floating point values.
@@ -46,7 +46,7 @@ func (noopMetricsProvider) Gauge(string, Tags) Gauge {
 
 type noopCounter struct{}
 
-func (noopCounter) Add(int64) {}
+func (noopCounter) Add(float64) {}
 
 type noopHistogram struct{}
